+++ rust/stackablectl/CHANGELOG.md
@@ -6,8 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- Support listing endpoints of Listeners in `stackablectl stacklet list`.
-  Currently only HDFS is using listener-op, so we can only test that so far ([#XXX]).
+- Support listing endpoints of Listeners in `stackablectl stacklet list` command.
+  Currently only HDFS is using listener-op, so we can only test that so far ([#213]).
 
 ### Changed
 
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file.
 
 [#181]: https://github.com/stackabletech/stackable-cockpit/pull/181
 [#202]: https://github.com/stackabletech/stackable-cockpit/pull/202
+[#213]: https://github.com/stackabletech/stackable-cockpit/pull/213
 
 ## [23.11.3] - 2024-01-03